MidiLogic added & MidiController changes
MIDIController -> MidiController
